backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.8
django-cors-headers>=3.8
djangorestframework>=3.14.0
drf-yasg>=1.21.7